requirements.txt
streamlit
opencv-python-headless 
pillow
facenet-pytorch
opencv-python
numpy

